fix: circular dependency
utils and const depend on each other.
Ref #294
